move to float

+++ schemas.py
@@ -116,8 +116,8 @@ class AccountDetails(BaseModel):
     new_stripe_subscription_id: Optional[str] = None
     new_subscription_id: Optional[int] = None
     test_results_used: int
-    build_credits_used_this_month: Optional[int] = None
-    build_credits_remaining_before_topup: Optional[int] = None
+    build_credits_used_this_month: Optional[float] = None
+    build_credits_remaining_before_topup: Optional[float] = None
     users: int
 
     payment_failed: Optional[bool] = None
